+++ Calculator/Calculator/Formula.swift
@@ -16,14 +16,3 @@ struct Formula {
         return result
     }
 }
-
-extension String {
-    func split(with target: Character) -> [String] {
-        let splitedString = self.split(separator: target).map { String($0) }
-        return splitedString
-    }
-}
-
-
-
-
