+++ Calculator/Calculator/extension.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+extension Double: CalculateItem {
+    
+}
+
+extension String {
+    func split(with target: Character) -> [String] {
+        let splitedString = self.split(separator: target).map { String($0) }
+        return splitedString
+    }
+}
